+++ src/main/java/server/tigglemate/domain/accountBook/domain/repository/AccountBookRepository.java
@@ -11,8 +11,8 @@ public interface AccountBookRepository extends JpaRepository<AccountBook, Long> 
     Optional<AccountBook> findById(int userId);
 
     @Query("SELECT a FROM AccountBook a WHERE a.id = :id")
-    AccountBook findAccountBookById(int id);
+    AccountBook findAccountBookById(Long id);
 
     @Query("SELECT a.targetAmount FROM AccountBook a WHERE a.id =:id")
-    Integer getTargetAmount(int id);
+    Integer getTargetAmount(Long id);
 }
